+++ fe/fe-core/src/main/java/com/starrocks/external/delta/DeltaUtils.java
@@ -3,11 +3,8 @@
 package com.starrocks.external.delta;
 
 import com.google.common.collect.Lists;
-import com.starrocks.catalog.ArrayType;
 import com.starrocks.catalog.Column;
 import com.starrocks.catalog.DeltaLakeTable;
-import com.starrocks.catalog.PrimitiveType;
-import com.starrocks.catalog.ScalarType;
 import com.starrocks.catalog.Type;
 import com.starrocks.external.hive.RemoteFileInputFormat;
 import com.starrocks.external.iceberg.StarRocksIcebergException;
@@ -20,6 +17,7 @@ import org.apache.hadoop.conf.Configuration;
 
 import java.util.List;
 
+import static com.starrocks.external.ColumnTypeConverter.fromDeltaLakeType;
 import static com.starrocks.external.hive.HiveMetastoreApiConverter.CONNECTOR_ID_GENERATOR;
 
 public class DeltaUtils {
@@ -43,7 +41,7 @@ public class DeltaUtils {
 
         for (StructField field : metadata.getSchema().getFields()) {
             DataType dataType = field.getDataType();
-            Type srType = convertColumnType(dataType);
+            Type srType = fromDeltaLakeType(dataType);
             Column column = new Column(field.getName(), srType, true);
             fullSchema.add(column);
         }
@@ -52,71 +50,7 @@ public class DeltaUtils {
                 fullSchema, metadata.getPartitionColumns(), deltaLog);
     }
 
-    public static Type convertColumnType(DataType dataType) {
-        if (dataType == null) {
-            return Type.NULL;
-        }
-        PrimitiveType primitiveType;
-        DeltaDataType deltaDataType = DeltaDataType.instanceFrom(dataType.getClass());
-        switch (deltaDataType) {
-            case BOOLEAN:
-                primitiveType = PrimitiveType.BOOLEAN;
-                break;
-            case BYTE:
-            case TINYINT:
-                primitiveType = PrimitiveType.TINYINT;
-                break;
-            case SMALLINT:
-                primitiveType = PrimitiveType.SMALLINT;
-                break;
-            case INTEGER:
-                primitiveType = PrimitiveType.INT;
-                break;
-            case LONG:
-                primitiveType = PrimitiveType.BIGINT;
-                break;
-            case FLOAT:
-                primitiveType = PrimitiveType.FLOAT;
-                break;
-            case DOUBLE:
-                primitiveType = PrimitiveType.DOUBLE;
-                break;
-            case DATE:
-                primitiveType = PrimitiveType.DATE;
-                break;
-            case TIMESTAMP:
-                primitiveType = PrimitiveType.DATETIME;
-                break;
-            case STRING:
-                return ScalarType.createDefaultString();
-            case DECIMAL:
-                int precision = ((io.delta.standalone.types.DecimalType) dataType).getPrecision();
-                int scale = ((io.delta.standalone.types.DecimalType) dataType).getScale();
-                return ScalarType.createUnifiedDecimalType(precision, scale);
-            case ARRAY:
-                Type type = convertToArrayType((io.delta.standalone.types.ArrayType) dataType);
-                if (type.isArrayType()) {
-                    return type;
-                } else {
-                    return Type.UNKNOWN_TYPE;
-                }
-            case NULL:
-                primitiveType = PrimitiveType.NULL_TYPE;
-                break;
-            case BINARY:
-            case MAP:
-            case STRUCT:
-            default:
-                primitiveType = PrimitiveType.UNKNOWN_TYPE;
-        }
-        return ScalarType.createType(primitiveType);
-    }
-
-    private static ArrayType convertToArrayType(io.delta.standalone.types.ArrayType arrayType) {
-        return new ArrayType(convertColumnType(arrayType.getElementType()));
-    }
-
-    public static RemoteFileInputFormat getHdfsFileFormat(String format) {
+    public static RemoteFileInputFormat getRemoteFileFormat(String format) {
         if (format.equalsIgnoreCase("ORC")) {
             return RemoteFileInputFormat.ORC;
         } else if (format.equalsIgnoreCase("PARQUET")) {
